Rearrange scan functions.
