--- internal/codeintel/stores/dbstore/indexes.go
@@ -1,45 +1,0 @@
-package dbstore
-
-import (
-	"github.com/lib/pq"
-
-	"github.com/sourcegraph/sourcegraph/internal/codeintel/types"
-	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
-	"github.com/sourcegraph/sourcegraph/internal/workerutil"
-	dbworkerstore "github.com/sourcegraph/sourcegraph/internal/workerutil/dbworker/store"
-)
-
-func scanIndex(s dbutil.Scanner) (index types.Index, err error) {
-	var executionLogs []dbworkerstore.ExecutionLogEntry
-	if err := s.Scan(
-		&index.ID,
-		&index.Commit,
-		&index.QueuedAt,
-		&index.State,
-		&index.FailureMessage,
-		&index.StartedAt,
-		&index.FinishedAt,
-		&index.ProcessAfter,
-		&index.NumResets,
-		&index.NumFailures,
-		&index.RepositoryID,
-		&index.RepositoryName,
-		pq.Array(&index.DockerSteps),
-		&index.Root,
-		&index.Indexer,
-		pq.Array(&index.IndexerArgs),
-		&index.Outfile,
-		pq.Array(&executionLogs),
-		&index.Rank,
-		pq.Array(&index.LocalSteps),
-		&index.AssociatedUploadID,
-	); err != nil {
-		return index, err
-	}
-
-	for _, entry := range executionLogs {
-		index.ExecutionLogs = append(index.ExecutionLogs, workerutil.ExecutionLogEntry(entry))
-	}
-
-	return index, nil
-}
